# packages/api-server/src/openapi/openapi.yaml
openapi: 3.0.0
info:
  title: ConfigBuddy CMDB API
  version: 2.0.0
  description: |
    Comprehensive REST API for ConfigBuddy - Enterprise Configuration Management Database (CMDB) platform.

    ConfigBuddy v2.0 features:
    - Graph-based CI relationships with Neo4j
    - 37 connectors (17 TypeScript + 20 JSON-only)
    - Unified credential system with affinity matching
    - Discovery agent architecture with smart routing
    - AI-powered anomaly detection and impact analysis
    - Identity resolution and reconciliation

    ## Authentication

    All API endpoints (except `/api/v1/auth/login`) require authentication via:
    - **JWT Token**: Include `Authorization: Bearer <token>` header
    - **API Key**: Include `X-API-Key: <key>` header

    ## Rate Limiting

    Rate limits apply per authentication token/API key:
    - **Standard**: 1000 requests/hour
    - **Burst**: 100 requests/minute

    Rate limit headers included in responses:
    - `X-RateLimit-Limit`: Maximum requests per window
    - `X-RateLimit-Remaining`: Remaining requests
    - `X-RateLimit-Reset`: Unix timestamp when limit resets

  contact:
    name: ConfigBuddy Support
    url: http://localhost:8080
    email: support@configbuddy.io
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
  - url: http://staging-api.configbuddy.io/api/v1
    description: Staging server
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /analytics/discovery-stats:
    get:
      tags: [Analytics]
      summary: Discovery statistics
      description: Get discovery job statistics
      operationId: getDiscoveryStats
      parameters:
        - name: _start_date
          in: query
          schema:
            type: string
            format: date-time
        - name: _end_date
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Discovery statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_jobs:
                    type: integer
                  successful_jobs:
                    type: integer
                  failed_jobs:
                    type: integer
                  cis_discovered:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'

  # =============================================================================
  # ANOMALIES
  # =============================================================================
...
  /jobs/stats:
    get:
      tags: [Jobs]
      summary: Get job statistics
      description: Overall job statistics across all queues
      operationId: getJobStats
      responses:
        '200':
          description: Job statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_jobs:
                    type: integer
                  active_jobs:
                    type: integer
                  completed_jobs:
                    type: integer
                  failed_jobs:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
  /jobs/discovery:
    get:
      tags: [Jobs]
      summary: List discovery jobs
      description: Get all discovery jobs
      operationId: listDiscoveryJobs
      responses:
        '200':
          description: Discovery jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Job'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /jobs/discovery/{provider}:
    post:
      tags: [Jobs]
      summary: Trigger discovery
      description: Manually trigger discovery for a provider
      operationId: triggerDiscovery
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/DiscoveryProvider'
      responses:
        '202':
          description: Discovery job queued
          content:
            application/json:
              schema:
                type: object
                properties:
                  job_id:
                    type: string
                  queue:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
  /jobs/{queueName}/{jobId}:
    get:
      tags: [Jobs]
      summary: Get job status
      description: Get detailed status of a specific job
      operationId: getJobStatus
      parameters:
        - name: queueName
          in: path
          required: true
          schema:
            type: string
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Job details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Jobs]
      summary: Cancel job
      description: Cancel a running or queued job
      operationId: cancelJob
      parameters:
        - name: queueName
          in: path
          required: true
          schema:
            type: string
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Job cancelled
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /jobs/{queueName}/{jobId}/retry:
    post:
      tags: [Jobs]
      summary: Retry failed job
      description: Retry a failed job
      operationId: retryJob
      parameters:
        - name: queueName
          in: path
          required: true
          schema:
            type: string
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      responses:
        '202':
          description: Job requeued
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # =============================================================================
  # HEALTH
  # =============================================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from /auth/login endpoint
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key from /auth/api-keys endpoint
...
  responses:
...
    Unauthorized:
      description: Authentication required or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Unauthorized
            message: Invalid or expired JWT token
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Not Found
            message: CI with ID 'srv-prod-web-01' not found
...
  schemas:
...
    DiscoveryProvider:
      type: string
      enum:
        - aws
        - azure
        - gcp
        - vmware
        - kubernetes
        - docker
        - nmap
        - ssh
        - snmp
        - active_directory
        - servicenow
        - salesforce
        - datadog
        - pagerduty
...
    Job:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        queue:
          type: string
        status:
          type: string
          enum: [waiting, active, completed, failed, delayed]
        progress:
          type: number
        data:
          type: object
        result:
          type: object
        error:
          type: object
        created_at:
          type: string
          format: date-time
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time

    # =========================================================================
    # AUDIT
    # =========================================================================
...
    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: Error type/code
          example: Validation Error
        message:
          type: string
          description: Human-readable error message
          example: Invalid CI type provided
        details:
          type: array
          description: Detailed validation errors
          items:
            type: object
            properties:
              field:
                type: string
              error:
                type: string